scripts/requirements.txt
httpx==0.26.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0
geopy==2.4.1
